sysdeps: add sys_sigaction() on linux and move signal.h to abi-bits
