+++ sysdeps/linux/generic/sysdeps.cpp
@@ -16,6 +16,7 @@
 #define NR_fstat 5
 #define NR_lseek 8
 #define NR_mmap 9
+#define NR_sigaction 13
 #define NR_ioctl 16
 #define NR_exit 60
 #define NR_arch_prctl 158
@@ -136,6 +137,14 @@ int sys_stat(fsfd_target fsfdt, int fd, const char *path, int flags, struct stat
                 ret = do_syscall(NR_fstat, fd, statbuf);
         else
                 ret = do_syscall(NR_stat, path, statbuf);
+        if (int e = sc_error(ret); e)
+                return e;
+        return 0;
+}
+
+int sys_sigaction(int signum, const struct sigaction *act,
+                struct sigaction *oldact) {
+        auto ret = do_syscall(NR_sigaction, signum, act, oldact, sizeof(sigset_t));
         if (int e = sc_error(ret); e)
                 return e;
         return 0;
