+++ sysdeps/linux/include/abi-bits/signal.h
@@ -1,0 +1,1 @@
+../../../../abis/linux/x86_64/signal.h
