+++ sysdeps/qword/include/abi-bits/signal.h
@@ -1,0 +1,1 @@
+../../../../abis/mlibc/signal.h
