+++ sysdeps/sigma/meson.build
@@ -26,6 +26,7 @@ if not no_headers
 		'include/abi-bits/vm-flags.h',
 		'include/abi-bits/errno.h',
 		'include/abi-bits/stat.h',
+		'include/abi-bits/signal.h',
 		subdir: 'abi-bits'
 	)
 endif
